+++ src/Util/Lifecycle/Update.php
@@ -155,6 +155,10 @@ class Update
         if (\version_compare($updateContext->getCurrentPluginVersion(), '9.10.0', '<')) {
             $this->updateTo9100($updateContext->getContext());
         }
+
+        if (\version_compare($updateContext->getCurrentPluginVersion(), '10.6.0', '<')) {
+            $this->updateTo1060();
+        }
     }
 
     private function updateTo130(): void
@@ -579,5 +583,10 @@ class Update
     private function updateTo9100(Context $context): void
     {
         $this->paymentMethodInstaller->updateAllMedia($context);
+    }
+
+    private function updateTo1060(): void
+    {
+        $this->systemConfig->set(Settings::ECS_SHIPPING_CALLBACK_ENABLED, true);
     }
 }