+++ tests/Checkout/ExpressCheckout/SalesChannel/ExpressCreateOrderRouteTest.php
@@ -16,6 +16,7 @@ use Shopware\Core\Checkout\Cart\LineItem\LineItem;
 use Shopware\Core\Checkout\Cart\SalesChannel\CartService;
 use Shopware\Core\Framework\Log\Package;
 use Shopware\Core\Framework\Test\TestCaseBase\IntegrationTestBehaviour;
+use Shopware\Core\PlatformRequest;
 use Shopware\Core\System\SystemConfig\SystemConfigService;
 use Swag\PayPal\Checkout\Cart\Service\CartPriceService;
 use Swag\PayPal\Checkout\Exception\OrderZeroValueException;
@@ -92,11 +93,11 @@ class ExpressCreateOrderRouteTest extends TestCase
         static::assertTrue($this->logger->hasDebug('Configured shipping callback'));
     }
 
-    public function testCreateWithoutShippingCallback(): void
+    public function testCreateWithLocalEnvironmentActive(): void
     {
         $salesChannelContext = $this->getSalesChannelContext();
 
-        $response = $this->createRoute(true)->createPayPalOrder(new Request(), $salesChannelContext);
+        $response = $this->createRoute([Settings::IS_LOCAL_ENVIRONMENT => true])->createPayPalOrder(new Request(), $salesChannelContext);
 
         static::assertSame(Response::HTTP_OK, $response->getStatusCode());
         static::assertSame(CreateOrderCapture::ID, $response->getToken());
@@ -104,12 +105,86 @@ class ExpressCreateOrderRouteTest extends TestCase
         static::assertTrue($this->logger->hasDebug('Skipped shipping callback due to being disabled in system config'));
     }
 
-    private function createRoute(bool $callbacksDisabled = false): ExpressCreateOrderRoute
+    public function testCreateWithShippingCallbackDisabled(): void
     {
+
+        $salesChannelContext = $this->getSalesChannelContext();
+
+        $response = $this->createRoute([Settings::ECS_SHIPPING_CALLBACK_ENABLED => false])->createPayPalOrder(new Request(), $salesChannelContext);
+
+        static::assertSame(Response::HTTP_OK, $response->getStatusCode());
+        static::assertSame(CreateOrderCapture::ID, $response->getToken());
+
+        static::assertTrue($this->logger->hasDebug('Skipped shipping callback due to being disabled in system config'));
+    }
+
+    public function testCreateShippingCallbackStoreApi(): void
+    {
+        $salesChannelContext = $this->getSalesChannelContext();
+
+        $router = $this->createMock(RouterInterface::class);
+        $router
+            ->expects($this->once())
+            ->method('generate')
+            ->with('store-api.paypal.express.shipping_callback')
+            ->willReturn('generatedUrl');
+
+        $response = $this->createRoute([], $router)->createPayPalOrder(new Request(), $salesChannelContext);
+
+        static::assertSame(Response::HTTP_OK, $response->getStatusCode());
+        static::assertSame(CreateOrderCapture::ID, $response->getToken());
+
+        $request = $this->getClient()->getLast();
+        static::assertNotNull($request);
+
+        $order = (new Order())->assign($request->getRequestBody() ?? []);
+
+        $experienceContext = $order->getPaymentSource()?->getPaypal()?->getExperienceContext();
+        static::assertNotNull($experienceContext);
+        static::assertNotNull($experienceContext->getOrderUpdateCallbackConfig());
+    }
+
+    public function testCreateShippingCallbackStorefront(): void
+    {
+        $salesChannelContext = $this->getSalesChannelContext();
+
+        $router = $this->createMock(RouterInterface::class);
+        $router
+            ->expects($this->once())
+            ->method('generate')
+            ->with('frontend.paypal.express.shipping_callback')
+            ->willReturn('generatedUrl');
+
+        $request = new Request();
+        $request->attributes->set(PlatformRequest::ATTRIBUTE_ROUTE_SCOPE, ['storefront']);
+
+        $response = $this->createRoute([], $router)->createPayPalOrder($request, $salesChannelContext);
+
+        static::assertSame(Response::HTTP_OK, $response->getStatusCode());
+        static::assertSame(CreateOrderCapture::ID, $response->getToken());
+
+        $request = $this->getClient()->getLast();
+        static::assertNotNull($request);
+
+        $order = (new Order())->assign($request->getRequestBody() ?? []);
+
+        $experienceContext = $order->getPaymentSource()?->getPaypal()?->getExperienceContext();
+        static::assertNotNull($experienceContext);
+        static::assertNotNull($experienceContext->getOrderUpdateCallbackConfig());
+    }
+
+    /**
+     * @param array<string, mixed> $systemConfigSettings
+     */
+    private function createRoute(
+        array $systemConfigSettings = [],
+        ?RouterInterface $router = null,
+    ): ExpressCreateOrderRoute {
         $systemConfig = $this->createSystemConfigServiceMock([
             Settings::CLIENT_ID => 'testClientId',
             Settings::CLIENT_SECRET => 'testClientSecret',
-            Settings::IS_LOCAL_ENVIRONMENT => $callbacksDisabled,
+            Settings::ECS_SHIPPING_CALLBACK_ENABLED => true,
+            ...$systemConfigSettings,
         ]);
 
         $priceFormatter = new PriceFormatter();