+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 9.12.0
+- Added setting to disable the shipping callback for express checkouts.
+- Fixes an issue, where the shiiping callback required the store-api to be exposed.
+
 # 9.11.0
 - Added Austria to the countries where Pay Later is available